backend/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
comm==0.2.2
contourpy==1.3.1
cramjam==2.9.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dotenv==0.9.9
duckdb==1.1.3
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.17.0
findspark==2.0.1
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.1
mizani==0.13.5
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.6
plotnine==0.14.5
polars==1.20.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.0
PyAthena==3.12.2
pybaseball==2.2.7
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyGithub==2.6.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-learn==1.7.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.5.1
six==1.17.0
sklearn-preprocessing==0.1.0
sniffio==1.3.1
soupsieve==2.6
sportypy==1.0.0
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==3.0.2
yarl==1.18.3
